chatbot/Part B/requirements.txt
google-generativeai
requests
beautifulsoup4
PyPDF2
python-dotenv